Target Sum in an Array
Targeted Sum
Given an array of numbers and an integer target, find two unique numbers in the array that add up to the target value. Return an array with the indices of those two numbers, or "Target not found" if no two numbers sum up to the target.

The returned array should have the indices in ascending order.
Tests
find_target([1, 3, 4, 6, 7, 8], 15) should return [4, 5].
find_target([1, 3, 4, 7], 14) should return 'Target not found'.